+++ drivers/ESP32/bdring_v4_map.h
@@ -98,8 +98,9 @@
 #if MODBUS_ENABLE
 #define UART2_RX_PIN            GPIO_NUM_22
 #define UART2_TX_PIN            GPIO_NUM_21
+#if RS485_DIR_ENABLE
 #define MODBUS_DIRECTION_PIN    GPIO_NUM_2
-#define MODBUS_BAUD             19200
+#endif
 #endif