+++ grbl/hal.h
@@ -353,6 +353,9 @@ typedef void (*on_program_completed_ptr)(program_flow_t program_flow);
 typedef void (*on_execute_realtime_ptr)(uint_fast16_t state);
 typedef void (*on_unknown_accessory_override_ptr)(uint8_t cmd);
 typedef void (*on_report_options_ptr)(void);
+typedef void (*on_report_command_help_ptr)(void);
+typedef setting_details_t *(*on_report_settings_ptr)(void); // NOTE: this must match the signature of the same definition in
+                                                            // the setting_details_t structure in settings.h!
 typedef void (*on_realtime_report_ptr)(stream_write_ptr stream_write, report_tracking_flags_t report);
 typedef void (*on_unknown_feedback_message_ptr)(stream_write_ptr stream_write);
 typedef bool (*on_laser_ppi_enable_ptr)(uint_fast16_t ppi, uint_fast16_t pulse_length);
@@ -369,6 +372,8 @@ typedef struct {
     on_execute_realtime_ptr on_execute_realtime;
     on_unknown_accessory_override_ptr on_unknown_accessory_override;
     on_report_options_ptr on_report_options;
+    on_report_command_help_ptr on_report_command_help;
+    on_report_settings_ptr on_report_settings;
     on_realtime_report_ptr on_realtime_report;
     on_unknown_feedback_message_ptr on_unknown_feedback_message;
     on_unknown_sys_command_ptr on_unknown_sys_command; // return Status_Unhandled if not handled.