+++ plugins/odometer/odometer.c
@@ -57,6 +57,7 @@ static on_state_change_ptr on_state_change;
 static spindle_set_state_ptr spindle_set_state_;
 static settings_changed_ptr settings_changed;
 static on_report_options_ptr on_report_options;
+static on_report_command_help_ptr on_report_command_help;
 
 static void stepperPulseStart (stepper_t *stepper)
 {
@@ -214,6 +215,16 @@ static status_code_t commandExecute (uint_fast16_t state, char *line, char *lcli
     return retval == Status_Unhandled && on_unknown_sys_command ? on_unknown_sys_command(state, line, lcline) : retval;
 }
 
+static void onReportCommandHelp (void)
+{
+    hal.stream.write("$ODOMETERS - list odometer log" ASCII_EOL);
+    hal.stream.write("$ODOMETERS=PREV - list previous odometer log when available" ASCII_EOL);
+    hal.stream.write("$ODOMETERS=RST - copy current log to previous and clear current" ASCII_EOL);
+
+    if(on_report_command_help)
+        on_report_command_help();
+}
+
 static void onReportOptions (void)
 {
     on_report_options();
@@ -254,6 +265,9 @@ void odometer_init()
 
         on_report_options = grbl.on_report_options;
         grbl.on_report_options = onReportOptions;
+
+        on_report_command_help = grbl.on_report_command_help;
+        grbl.on_report_command_help = onReportCommandHelp;
 
         settings_changed = hal.settings_changed;
         hal.settings_changed = onSettingsChanged;