+++ plugins/spindle/modbus.c
@@ -2,7 +2,7 @@
 
   modbus.c - a lightweigth ModBus implementation
 
-  Part of GrblHAL
+  Part of grblHAL
 
   Copyright (c) 2020 Terje Io
   except modbus_CRC16x which is Copyright (c) 2006 Christian Walter <wolti@sil.at>
@@ -28,12 +28,17 @@
 #ifdef ARDUINO
 #include "../grbl/hal.h"
 #include "../grbl/protocol.h"
+#include "../grbl/nvs_buffer.h"
 #else
 #include "grbl/hal.h"
 #include "grbl/protocol.h"
+#include "grbl/settings.h"
+#include "grbl/nvs_buffer.h"
 #endif
 
 #include "modbus.h"
+
+#define DEFAULT_BAUDRATE 3 // 19200
 
 typedef struct queue_entry {
     bool async;
@@ -42,16 +47,22 @@ typedef struct queue_entry {
     struct queue_entry *next;
 } queue_entry_t;
 
+static const uint32_t baud[]    = { 2400, 4800, 9600, 19200, 38400, 115200 };
+static const uint16_t silence[] = {   16,    8,    4,     2,     2,      2 };
+
 static modbus_stream_t *stream;
-static uint16_t rx_timeout = 0;
+static uint16_t rx_timeout = 0, silence_until = 0, silence_timeout;
 static int16_t exception_code = 0;
 static queue_entry_t queue[MODBUS_QUEUE_LENGTH];
+static modbus_settings_t modbus;
 static volatile bool spin_lock = false;
 static volatile queue_entry_t *tail, *head, *packet = NULL;
 static volatile modbus_state_t state = ModBus_Idle;
+
 static driver_reset_ptr driver_reset;
 static on_execute_realtime_ptr on_execute_realtime;
 static on_report_options_ptr on_report_options;
+static driver_setting_ptrs_t driver_settings;
 
 // Compute the MODBUS RTU CRC
 static uint16_t modbus_CRC16x (char *buf, uint_fast16_t len)
@@ -73,14 +84,14 @@ static uint16_t modbus_CRC16x (char *buf, uint_fast16_t len)
     // Note, this number has low and high bytes swapped, so use it accordingly (or swap bytes)
     return crc;
 }
-
+/*
 static bool valid_crc (char *buf, uint_fast16_t len)
 {
     uint16_t crc = modbus_CRC16x(buf, len - 2);
 
     return buf[len - 1] == (crc >> 8) && buf[len - 2] == (crc & 0xFF);
 }
-
+*/
 void modbus_poll (uint_fast16_t grbl_state)
 {
     static uint32_t last_ms;
@@ -104,7 +115,7 @@ void modbus_poll (uint_fast16_t grbl_state)
                 packet = tail;
                 tail = tail->next;
                 state = ModBus_TX;
-                rx_timeout = stream->rx_timeout;
+                rx_timeout = modbus.rx_timeout;
 
                 if(stream->set_direction)
                     stream->set_direction(true);
@@ -112,6 +123,13 @@ void modbus_poll (uint_fast16_t grbl_state)
                 packet->sent = true;
                 stream->flush_rx_buffer();
                 stream->write(((queue_entry_t *)packet)->msg.adu, ((queue_entry_t *)packet)->msg.tx_length);
+            }
+            break;
+
+        case ModBus_Silent:
+            if(ms >= silence_until) {
+                silence_until = 0;
+                state = ModBus_Idle;
             }
             break;
 
@@ -128,7 +146,7 @@ void modbus_poll (uint_fast16_t grbl_state)
         case ModBus_AwaitReply:
             if(rx_timeout && --rx_timeout == 0) {
                 if(packet->async)
-                    state = ModBus_Idle;
+                    state = ModBus_Silent;
                 else if(stream->read() == 1 && (stream->read() & 0x80)) {
                     exception_code = stream->read();
                     state = ModBus_Exception;
@@ -136,6 +154,8 @@ void modbus_poll (uint_fast16_t grbl_state)
                     state = ModBus_Timeout;
                 packet = NULL;
                 spin_lock = false;
+                if(state != ModBus_AwaitReply)
+                    silence_until = ms + silence_timeout;
                 return;
             }
 
@@ -147,11 +167,13 @@ void modbus_poll (uint_fast16_t grbl_state)
                     *buf++ = stream->read();
                 } while(--packet->msg.rx_length);
 
-                if((state = packet->async ? ModBus_Idle : ModBus_GotReply) == ModBus_Idle)
+                if((state = packet->async ? ModBus_Silent : ModBus_GotReply) == ModBus_Silent)
                     stream->on_rx_packet(&((queue_entry_t *)packet)->msg);
 
+                silence_until = ms + silence_timeout;
                 packet = NULL;
             }
+
             break;
 
         default:
@@ -178,7 +200,8 @@ bool modbus_send (modbus_message_t *msg, bool block)
         bool poll = true;
 
         while(state != ModBus_Idle) {
-            if(!protocol_execute_realtime())
+            grbl.on_execute_realtime(sys.state);
+            if(ABORTED)
                 return false;
         }
 
@@ -186,7 +209,7 @@ bool modbus_send (modbus_message_t *msg, bool block)
             stream->set_direction(true);
 
         state = ModBus_TX;
-        rx_timeout = stream->rx_timeout;
+        rx_timeout = modbus.rx_timeout;
 
         memcpy(&sync_msg.msg, msg, sizeof(modbus_message_t));
 
@@ -198,7 +221,9 @@ bool modbus_send (modbus_message_t *msg, bool block)
 
         while(poll) {
 
-            if(!protocol_execute_realtime())
+            grbl.on_execute_realtime(sys.state);
+
+            if(ABORTED)
                 poll = false;
 
             else switch(state) {
@@ -224,7 +249,7 @@ bool modbus_send (modbus_message_t *msg, bool block)
             }
         }
 
-        state = ModBus_Idle;
+        state = silence_until > 0 ? ModBus_Silent : ModBus_Idle;
 
     } else if(packet != &sync_msg) {
         if(head->next != tail) {
@@ -249,6 +274,8 @@ static void modbus_reset (void)
 
     packet = NULL;
     tail = head;
+
+    silence_until = 0;
     state = ModBus_Idle;
 
     stream->flush_tx_buffer();
@@ -257,32 +284,158 @@ static void modbus_reset (void)
     driver_reset();
 }
 
+static uint32_t get_baudrate (uint32_t rate)
+{
+    uint32_t idx = sizeof(baud) / sizeof(uint32_t);
+
+    do {
+        if(baud[--idx] == rate)
+            return idx;
+    } while(idx);
+
+    return DEFAULT_BAUDRATE;
+}
+
+static const setting_group_detail_t modbus_groups [] = {
+    { Group_Root, Group_ModBus, "ModBus"}
+};
+
+static const setting_detail_t modbus_settings[] = {
+    { Settings_ModBus_BaudRate, Group_ModBus, "ModBus baud rate", NULL, Format_RadioButtons, "2400,4800,9600,19200,38400,115200", NULL, NULL },
+    { Settings_ModBus_RXTimeout, Group_ModBus, "ModBus RX timeout", "milliseconds", Format_Integer, "####0", "50", "250" }
+};
+
+static setting_details_t details = {
+    .groups = modbus_groups,
+    .n_groups = sizeof(modbus_groups) / sizeof(setting_group_detail_t),
+    .settings = modbus_settings,
+    .n_settings = sizeof(modbus_settings) / sizeof(setting_detail_t)
+};
+
+setting_details_t *on_report_settings (void)
+{
+    return &details;
+}
+
+static status_code_t modbus_setting (setting_type_t setting, float value, char *svalue)
+{
+    status_code_t status = svalue ? Status_OK : Status_Unhandled;
+
+    if(svalue) switch(setting) {
+
+        case Settings_ModBus_BaudRate:
+            if(isintf(value) && value >= 0.0f && value < sizeof(baud) / sizeof(uint32_t)) {
+                modbus.baud_rate = baud[(uint32_t)value];
+                silence_timeout = silence[(uint32_t)value];
+                stream->set_baud_rate(modbus.baud_rate);
+            } else
+                status = Status_InvalidStatement; //out of range...
+            break;
+
+        case Settings_ModBus_RXTimeout:
+            if(isintf(value) && value >= 50.0f && value < 250.0f) {
+                modbus.rx_timeout = (uint32_t)value;
+            } else
+                status = Status_InvalidStatement; //out of range...
+            break;
+
+        default:
+            status = Status_Unhandled;
+            break;
+    }
+
+    if(status == Status_OK)
+        hal.nvs.memcpy_to_nvs(driver_settings.nvs_address, (uint8_t *)&modbus, sizeof(modbus_settings_t), true);
+
+    return status == Status_Unhandled && driver_settings.set ? driver_settings.set(setting, value, svalue) : status;
+}
+
+static void modbus_settings_report (setting_type_t setting)
+{
+    bool reported = true;
+
+    switch(setting) {
+
+        case Settings_ModBus_BaudRate:
+            report_uint_setting(setting, get_baudrate(modbus.baud_rate));
+            break;
+
+        case Settings_ModBus_RXTimeout:
+            report_uint_setting(setting, modbus.rx_timeout);
+            break;
+
+        default:
+            reported = false;
+            break;
+    }
+
+    if(!reported && driver_settings.report)
+        driver_settings.report(setting);
+}
+
+static void modbus_settings_restore (void)
+{
+    modbus.rx_timeout = 50;
+    modbus.baud_rate = baud[DEFAULT_BAUDRATE];
+
+    hal.nvs.memcpy_to_nvs(driver_settings.nvs_address, (uint8_t *)&modbus, sizeof(modbus_settings_t), true);
+
+    if(driver_settings.restore)
+        driver_settings.restore();
+}
+
+static void modbus_settings_load (void)
+{
+    if(hal.nvs.memcpy_from_nvs((uint8_t *)&modbus, driver_settings.nvs_address, sizeof(modbus_settings_t), true) != NVS_TransferResult_OK)
+        modbus_settings_restore();
+
+    silence_timeout = silence[get_baudrate(modbus.baud_rate)];
+
+    stream->set_baud_rate(modbus.baud_rate);
+
+    if(driver_settings.load)
+        driver_settings.load();
+}
+
 static void onReportOptions (void)
 {
     on_report_options();
-    hal.stream.write("[PLUGIN:MODBUS v0.01]" ASCII_EOL);
+    hal.stream.write("[PLUGIN:MODBUS v0.02]" ASCII_EOL);
 }
 
-void modbus_init (modbus_stream_t *mstream)
+bool modbus_init (modbus_stream_t *mstream)
 {
-    uint_fast8_t idx;
+    if((hal.driver_settings.nvs_address = nvs_alloc(sizeof(modbus_settings_t)))) {
+        memcpy(&driver_settings, &hal.driver_settings, sizeof(driver_setting_ptrs_t));
 
-    stream = mstream;
+        hal.driver_settings.set = modbus_setting;
+        hal.driver_settings.report = modbus_settings_report;
+        hal.driver_settings.load = modbus_settings_load;
+        hal.driver_settings.restore = modbus_settings_restore;
 
-    if(driver_reset == NULL) {
+        stream = mstream;
 
-        driver_reset = hal.driver_reset;
-        hal.driver_reset = modbus_reset;
+        if(driver_reset == NULL) {
 
-        on_execute_realtime = grbl.on_execute_realtime;
-        grbl.on_execute_realtime = modbus_poll;
+            driver_reset = hal.driver_reset;
+            hal.driver_reset = modbus_reset;
 
-        on_report_options = grbl.on_report_options;
-        grbl.on_report_options = onReportOptions;
+            on_execute_realtime = grbl.on_execute_realtime;
+            grbl.on_execute_realtime = modbus_poll;
+
+            on_report_options = grbl.on_report_options;
+            grbl.on_report_options = onReportOptions;
+
+            details.on_report_settings = grbl.on_report_settings;
+            grbl.on_report_settings = on_report_settings;
+        }
+
+        head = tail = &queue[0];
+
+        uint_fast8_t idx;
+        for(idx = 0; idx < MODBUS_QUEUE_LENGTH; idx++)
+            queue[idx].next = idx == MODBUS_QUEUE_LENGTH - 1 ? &queue[0] : &queue[idx + 1];
     }
 
-    head = tail = &queue[0];
-
-    for(idx = 0; idx < MODBUS_QUEUE_LENGTH; idx++)
-        queue[idx].next = idx == MODBUS_QUEUE_LENGTH - 1 ? &queue[0] : &queue[idx + 1];
+    return driver_settings.nvs_address != 0;
 }